requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.3
numpy==1.23.1
oauthlib==3.2.0
opencv-python==4.6.0.66
pandas==1.4.3
protobuf==4.21.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygsheets==2.0.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tornado==6.2
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.11
